Refactor BasicVerifier to be package private instead of public.
